+++ HelpConsole/src/Beep.cs
@@ -6,7 +6,7 @@ namespace HelpConsole
     /// <summary>
     /// Beep.
     /// </summary>
-    public static partial class HelpConsole
+    public partial class HelpConsole
     {
         /// <summary>
         /// Beep for warning.
@@ -51,24 +51,41 @@ namespace HelpConsole
             if (melody == null)
             {
                 // Throwing null exception to indicate provided value is null.
-                throw new ArgumentNullException("Melody is null");
+                throw new ArgumentNullException(nameof(melody), "melody is null.");
+            }
+
+            // Checking if melody's notes is null.
+            if (melody.Notes == null)
+            {
+                // Throwing null exception to indicate provided value is null.
+                throw new ArgumentNullException(nameof(melody), "melody notes is null.");
             }
 
             // Checking if melody is null.
             if (melody.Notes.Length == 0)
             {
                 // Throwing null exception to indicate provided value is null.
-                throw new ArgumentNullException("Melody content is null");
+                throw new ArgumentNullException(nameof(melody), "melody doesn't have notes.");
             }
 
             // Loop for notes array.
             for (int i = 0; i < melody.Notes.Length; i = i + 3)
             {
-                // Checking if frequency of note is in range, it
-                if (melody.Notes[i] < 37 || melody.Notes[i] > 32767 || melody.Notes[i + 1] < 0 || melody.Notes[i + 2] < 0)
+                // Checking if the frequency is in range. [frequency-beeping duration-waiting duration]
+                if (melody.Notes[i] < 37 || melody.Notes[i] > 32767)
                 {
                     //
-                    Console.WriteLine("Error on notes");
+                    WriteLine($"Error on notes. Note at index: {i} {(melody.Notes[i] < 37 ? "should be greater than 37" : "")} {(melody.Notes[i] > 32767 ? "should be lower than 32767" : "")}");
+
+                    //
+                    continue;
+                }
+
+                // Checking if the beeping duration or waitin duration are positive values.
+                if (melody.Notes[i + 1] < 0 || melody.Notes[i + 2] < 0)
+                {
+                    //
+                    WriteLine($"Error on notes. Notes at index: {i+1}/{i+2} {(melody.Notes[i + 1] < 0 ? "beeping duration should be greater than 0" : "")} {(melody.Notes[i + 2] < 0 ? "waiting duration should be greater than 0" : "")}");
 
                     //
                     continue;