+++ HelpConsole/src/Title.cs
@@ -5,19 +5,19 @@ namespace HelpConsole
     /// <summary>
     /// Console methods
     /// </summary>
-    public static partial class HelpConsole
+    public partial class HelpConsole
     {
         // Title's first part text.
-        private static string _titleFirstText = "";
+        private static string s_titleFirstText = "";
 
         // Title's second part text.
-        private static string _titleSecondText = "";
+        private static string s_titleSecondText = "";
 
         // Title's third part text.
-        private static string _titleThirdText = "";
+        private static string s_titleThirdText = "";
 
         // Title's seperator between first, second and third text.
-        private static string _titleSeperator = " - ";
+        private static string s_titleSeperator = " - ";
 
         /// <summary>
         /// Sets console's title seperator for better liability between parameters.
@@ -26,7 +26,7 @@ namespace HelpConsole
         public static void SetConsoleTitleSeperator(string seperator)
         {
             // Sets default variable with provided value.
-            _titleSeperator = seperator;
+            s_titleSeperator = seperator;
         }
 
         /// <summary>
@@ -55,7 +55,7 @@ namespace HelpConsole
         /// </summary>
         /// <param name="firstText">First text</param>
         /// <param name="secondText">Second text</param>
-        /// <param name="thirdText">Third text</param>
+        /// <param name="thirdText">Third text</param>        
         //[SupportedOSPlatform("windows")]
         public static void SetConsoleTitle(string firstText = "", string secondText = "", string thirdText = "")
         {
@@ -63,40 +63,40 @@ namespace HelpConsole
             if (firstText == null)
             {
                 // Set first text with default value.
-                firstText = _titleFirstText;
+                firstText = s_titleFirstText;
             }
             else
             {
                 // Set default value with given parameter.
-                _titleFirstText = firstText;
+                s_titleFirstText = firstText;
             }
 
             // Checking if secondText is null.
             if (secondText == null)
             {
                 // Set second text with default value.
-                secondText = _titleSecondText;
+                secondText = s_titleSecondText;
             }
             else
             {
                 // Set default value with given parameter.
-                _titleSecondText = secondText;
+                s_titleSecondText = secondText;
             }
 
             // Checking if thirdText is null.
             if (thirdText == null)
             {
                 // Set third text with default value.
-                thirdText = _titleThirdText;
+                thirdText = s_titleThirdText;
             }
             else
             {
                 // Set default value with given parameter.
-                _titleThirdText = thirdText;
+                s_titleThirdText = thirdText;
             }
 
             // Setting console's title with first, second and third text. Adds provided title seperator if parameter parameter is provided.
-            Console.Title = $"{firstText}{(string.IsNullOrEmpty(secondText) ? "" : _titleSeperator)}{secondText}{(string.IsNullOrEmpty(thirdText) ? "" : _titleSeperator)}{thirdText}";
+            Console.Title = $"{firstText}{(string.IsNullOrEmpty(secondText) ? "" : s_titleSeperator)}{secondText}{(string.IsNullOrEmpty(thirdText) ? "" : s_titleSeperator)}{thirdText}";
         }
 
         /// <summary>
@@ -106,13 +106,13 @@ namespace HelpConsole
         public static void ClearConsoleTitle()
         {
             // Set default value null.
-            _titleFirstText = null;
+            s_titleFirstText = null;
 
             // Set default value null.
-            _titleSecondText = null;
+            s_titleSecondText = null;
 
             // Set default value null.
-            _titleThirdText = null;
+            s_titleThirdText = null;
 
             // Set console's title empty.
             Console.Title = "";